moved models into their own package, also modified removal to search for id to simplify request body
