+++ .gitignore
@@ -7,3 +7,6 @@
 
 # Log files
 *.log
+
+# AWS Files
+.aws-sam/
