+++ src/main/java/com/revature/delete_user_favorites/UserFavoritesRepository.java
@@ -1,21 +1,15 @@
 package com.revature.delete_user_favorites;
 
 import com.revature.delete_user_favorites.exceptions.ResourceNotFoundException;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import com.revature.delete_user_favorites.models.User;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
 import software.amazon.awssdk.enhanced.dynamodb.Expression;
 import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
-import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean;
-import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbPartitionKey;
 import software.amazon.awssdk.enhanced.dynamodb.model.ScanEnhancedRequest;
 import software.amazon.awssdk.http.apache.ApacheHttpClient;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
-import java.util.List;
 
 /**
  * A repository-layer class used for querying DynamoDB using the DynamoDBMapper.
@@ -55,65 +49,5 @@ public class UserFavoritesRepository {
     public User saveUser(User user) {
         return userTable.updateItem(user);
     }
-}
-
-/**
- * The User POJO is necessary for storing the data received from DynamoDB.
- * It is very much a Data Transfer Object.
- */
-@Data
-@Builder
-@DynamoDbBean
-@NoArgsConstructor
-@AllArgsConstructor
-class User {
-    private String id;
-    private String username;
-    private List<SetDocument> favoriteSets;
-    private List<SetDocument> createdSets;
-    private String profilePicture;
-    private int points;
-    private int wins;
-    private int losses;
-    private String registrationDate;
-    private List<String> gameRecords;
-
-    @DynamoDbPartitionKey
-    public String getId() {
-        return id;
-    }
-}
-
-/**
- * SetDocument is necessary because the User returned from DynamoDB has a
- * list of SetDocuments attached to them describing the user's favorite cards
- * and their created list of cards. This is what we are altering.
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-class SetDocument {
-    private String id;
-    private String setName;
-    private List<Tags> tags;
-    private String author;
-    private boolean isPublic;
-    private int views;
-    private int plays;
-    private int studies;
-    private int favorites;
-}
-
-/**
- * Tags is a necessary pojo for persisting data from the DynamoDB to access
- * the users from the database. Without the Tags, it is impossible to handle
- * user data safely.
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-class Tags {
-    private String tagName;
-    private String tagColor;
 }
